backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
langchain==0.1.0
langchain-groq>=0.1.0
openai==1.3.7
groq>=0.10.0
websockets==12.0

